+++ R/cgal_version.R
@@ -4,12 +4,27 @@
 #' @export
 cgal_version <- function() {
   
-  # pkg_path = dirname(system.file(".", package = "RcppCGAL"))
-  # buildnumFile <- file.path(pkg_path, "VERSION")
-  # version <- readLines(buildnumFile)
-  version <- cgal_pkg_state$VERSION
-  message(version)
+  vers <- .cgal_version_check()
+  if (is.na(vers)) vers <- "Version not set yet. Likely you need to download the header files with cgal_install()."
+  message(vers)
 }
+
+.cgal_version_check <- function() {
+  vers <- cgal_pkg_state$VERSION
+  if (is.na(vers)) {
+    pkg_path <- .rcppcgal_package_path()
+    buildnumFile <- file.path(pkg_path, "VERSION")
+    
+    # only update version variable if file exists
+    if (file.exists(buildnumFile)) {
+      vers <- readLines(buildnumFile)
+      cgal_pkg_state$VERSION <- vers
+    } 
+    
+  }
+  return(cgal_pkg_state$VERSION)
+}
+
 
 #' Function to write CGAL version
 #'
@@ -20,18 +35,19 @@ cgal_version <- function() {
 #'
 #' @keywords internal
 .save_cgal_version <- function(version, own = FALSE) {
-  # pkg_path = dirname(system.file(".", package = "RcppCGAL"))
   
-  #save version number
-  # buildnumFile <- file.path(pkg_path, "VERSION")
+  # save old version
   cgal_pkg_state$OLD_VERSION <- cgal_pkg_state$VERSION
   
+  # create environmental variable
   if(!own) {
-    # writeLines(text = paste0("This is CGAL version ", version,"\n"), buildnumFile)
     cgal_pkg_state$VERSION <- paste0("This is CGAL version ", version)
   } else {
-    # writeLines(text = paste0("Supplied own CGAL from ",version,"\n"), buildnumFile)
-    cgal_pkg_stage$VERSION <- paste0("Supplied own CGAL from ", version)
+    cgal_pkg_state$VERSION <- paste0("Supplied own CGAL from ", version)
   }
   
+  #save version number
+  pkg_path <- .rcppcgal_package_path()
+  buildnumFile <- file.path(pkg_path, "VERSION")
+  writeLines(text =  cgal_pkg_state$VERSION, buildnumFile)
 }